+++ app/controllers/bookmark_lists_controller.rb
@@ -4,39 +4,6 @@ class BookmarkListsController < InheritedResources::Base
   before_action :authenticate_user!
   before_action :set_menu
 
-  def show
-    @bookmark_list = current_user.bookmark_lists.find(params[:id])
-
-    @page_title = Bookmark.model_name.human(count: 2)
-    @active_dashboard_menu = :bookmarks
-
-    all_bookmarks = current_user.bookmarks
-                                .includes({ item: [{ sub_categories: [:category] }, :brand] })
-                                .order(:created_at)
-
-    bookmarks = all_bookmarks.select { |bookmark| bookmark.bookmark_list_id == @bookmark_list.id }
-                             .map { |bookmark| BookmarkPresenter.new(bookmark) }
-
-    if params[:category].present?
-      sub_cat = SubCategory.friendly.find(params[:category])
-
-      if sub_cat
-        bookmarks =
-          bookmarks.select do |bookmark|
-            bookmark.product.sub_categories.include?(sub_cat)
-          end
-        @sub_category = sub_cat
-      end
-    end
-
-    @all_bookmarks = all_bookmarks.map { |bookmark| BookmarkPresenter.new(bookmark) }
-    @bookmarks = bookmarks
-    @bookmark_lists = current_user.bookmark_lists
-    @custom_attributes_for_products = CustomAttribute.all
-
-    @categories = get_grouped_sub_categories(bookmarks: @all_bookmarks)
-  end
-
   def new
     @bookmark_list = BookmarkList.new
 
@@ -61,7 +28,6 @@ class BookmarkListsController < InheritedResources::Base
       redirect_to dashboard_bookmark_list_path(@bookmark_list)
     else
       @active_dashboard_menu = :bookmarks
-      @bookmark_lists = current_user.bookmark_lists.order('LOWER(name)')
       render :new, status: :unprocessable_content
     end
   end